% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_artdata.R
\name{read_artdata}
\alias{read_artdata}
\title{Reads and combines multiple .txt files, retaining the source of each file}
\usage{
read_artdata(archivos, carpeta, sep = ",", header = TRUE)
}
\arguments{
\item{archivos}{A vector of .txt file names (e.g., "archivo1.txt", "archivo2.txt", ...).}

\item{carpeta}{The folder path where the files are located.}

\item{sep}{The separator for .txt files (e.g., \code{","} for comma, \code{";"} for semicolon, or \code{" "} for space).}

\item{header}{Logical, TRUE if the files have a header.}
}
\value{
A data frame with all files combined, including a column "archivo_origen" that indicates
the source file for each row.
}
\description{
This function reads multiple .txt files located in a specific folder, adds a column
indicating the source file, and then combines them into a single data frame.
}
\examples{
# Example for reading multiple .txt files
archivos_txt <- c("archivo1.txt", "archivo2.txt", "archivo3.txt")
carpeta <- "ruta/a/tu/carpeta"
if (all(file.exists(file.path(carpeta, archivos_txt)))) {
  datos_txt <- read_artdata_txt(archivos_txt, carpeta, sep = "\t", header = TRUE)
  print(datos_txt)
} else {
  warning("One or more .txt files are not found in the specified folder.")
}
}
