% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_artdata_2.R
\name{read_artdata_2}
\alias{read_artdata_2}
\title{Reads and combines multiple .csv files, retaining the source of each file}
\usage{
read_artdata_2(archivos, carpeta, sep = ",", header = TRUE)
}
\arguments{
\item{archivos}{A vector of .csv file names (e.g., "archivo1.csv", "archivo2.csv", ...).}

\item{carpeta}{The folder path where the files are located.}

\item{sep}{The separator for .csv files (e.g., \code{","} for comma, \code{";"} for semicolon).}

\item{header}{Logical, TRUE if the files have a header.}
}
\value{
A data frame with all files combined, including a column "archivo_origen" that indicates
the source file for each row.
}
\description{
This function reads multiple .csv files located in a specific folder, adds a column
indicating the source file, and then combines them into a single data frame.
}
\examples{
# Example for reading multiple .csv files
archivos_csv <- c("archivo1.csv", "archivo2.csv", "archivo3.csv")
carpeta <- "ruta/a/tu/carpeta"
if (all(file.exists(file.path(carpeta, archivos_csv)))) {
  datos_csv <- read_artdata_csv_2(archivos_csv, carpeta, sep = ",", header = TRUE)
  print(datos_csv)
} else {
  warning("One or more .csv files are not found in the specified folder.")
}
}
