% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remo_dup.R
\name{remo_dup}
\alias{remo_dup}
\title{Remove duplicates from a data frame}
\usage{
remo_dup(datos, mantener_ultima = FALSE)
}
\arguments{
\item{datos}{A data frame in which duplicates will be searched (for example, \code{artdata}).}

\item{mantener_ultima}{Logical, \code{TRUE} to keep the last occurrence of each duplicated row.
Defaults to \code{FALSE}, which keeps the first occurrence.}
}
\value{
A data frame without duplicate rows.
}
\description{
This function identifies and removes duplicate rows in a data frame that commonly occur in this type of records.
You can choose to remove duplicates while keeping only the last occurrence of each duplicated row.
}
\examples{
artdata_without_duplicates <- remo_dup(artdata, mantener_ultima = TRUE)
}
