% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAbarrida.R
\name{SAbarrida}
\alias{SAbarrida}
\title{Calculates the Swept Area (SA) based on distance and the width of the fishing gear.}
\usage{
SAbarrida(distancia, ancho)
}
\arguments{
\item{distancia}{A vector representing distances in meters.}

\item{ancho}{A vector representing the width of the fishing gear in meters.
It must be the same size as 'distancia' or a single value.}
}
\value{
A vector with the calculated swept area.
}
\description{
Calculates the Swept Area (SA) based on distance and the width of the fishing gear.
}
\examples{
# Calculate the swept area with a width of 2.5 meters
distancia <- c(100, 200, 300)
swept_area <- SAbarrida(distancia, ancho = 2.5)
swept_area_pre <- SAbarrida(distancia, ancho = 2.5)
}
