% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distart.R
\name{distart}
\alias{distart}
\title{Calcula la distancia euclideana en metros con información de los cuatro registros de latitud peor con diferentes tiempos}
\usage{
distart(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{point 1 Latitud}

\item{lon1}{point 1 Longitud}

\item{lat2}{point 2 Latitud}

\item{lon2}{point 2 Longitud}
}
\value{
distancia calculada
}
\description{
Calcula la distancia euclideana en metros con información de los cuatro registros de latitud peor con diferentes tiempos
}
\examples{
# Ejemplo de uso:
# Distancia entre dos puntos: (lat1, lon1) y (lat2, lon2)
distancia <- distart(
  40.7128, -74.0060,
  34.0522, -118.2437
)  # Distancia entre Nueva York y Los Ángeles
print(distancia)
}
