% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_artdata.R
\name{read_artdata}
\alias{read_artdata}
\title{Reads and combines multiple data files, retaining the source of each file}
\usage{
read_artdata(archivos, carpeta, sep = ";", header = TRUE)
}
\arguments{
\item{archivos}{Un vector de nombres de archivos (por ejemplo, "archivo1.txt", "archivo2.txt", ...).}

\item{carpeta}{La ruta de la carpeta donde se encuentran los archivos.}

\item{sep}{El separador utilizado en los archivos de texto (por defecto es ";").}

\item{header}{Lógico, TRUE si los archivos tienen encabezado.}
}
\value{
Un data frame con todos los archivos combinados, incluyendo una columna "archivo_origen" que indica
el archivo de origen para cada fila.
}
\description{
This function reads multiple text files located in a specific folder, adds a column
indicating the source file, and then combines them into a single data frame.
}
\examples{
archivos <- c("Draga_01.txt", "Draga_02.txt", "Draga_03.txt")
carpeta <- "ruta/a/tu/carpeta"
# Verificar que todos los archivos existen antes de ejecutar la función
if (all(file.exists(file.path(carpeta, archivos)))) {
  datos_combinados <- read_artdata(archivos, carpeta, sep = ",")
  print(datos_combinados)
} else {
  warning("Uno o más archivos no se encuentran en la carpeta especificada.")
}
}
