% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remo_dup.R
\name{remo_dup}
\alias{remo_dup}
\title{Elimina duplicados de un data frame}
\usage{
remo_dup(datos, mantener_ultima = FALSE)
}
\arguments{
\item{datos}{Un data frame en el que se buscarán duplicados (por ejemplo, \code{artdata}).}

\item{mantener_ultima}{Lógico, \code{TRUE} para mantener la última ocurrencia de cada fila duplicada.
Por defecto es \code{FALSE}, lo que mantiene la primera ocurrencia.}
}
\value{
Un data frame sin filas duplicadas.
}
\description{
Esta función identifica y elimina filas duplicadas en un data frame que comunente aparecen en este tipo de registros. Se puede optar por eliminar
los duplicados manteniendo solo la última ocurrencia de cada fila duplicada.
}
\examples{
artdata_sin_duplicados <- remo_dup(artdata, mantener_ultima = TRUE)
}
