% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAbarrida.R
\name{SAbarrida}
\alias{SAbarrida}
\title{Calcula el área barrida (SA) en función de la distancia y el ancho del arte de pesca.}
\usage{
SAbarrida(distancia, ancho)
}
\arguments{
\item{distancia}{Un vector que representa las distancias en metros.}

\item{ancho}{Un vector que representa el ancho del arte de pesca en metros.
Debe ser del mismo tamaño que 'distancia' o un solo valor.}
}
\value{
Un vector con el área barrida calculada.
}
\description{
Calcula el área barrida (SA) en función de la distancia y el ancho del arte de pesca.
}
\examples{
# Cálculo del área barrida con un ancho de 2.5 metros
distancias <- c(100, 200, 300)
area_barrida <- SAbarrida(distancias, ancho = 2.5)
area_barrida_pre <- SAbarrida(distancias, ancho = 2.5)
}
